@@ -1,20 +1,20 @@
 name: stackdriver_dart
 description: logging stackdriver report
-version: 1.0.7
+version: 1.1.0
 homepage: https://playon.jp
 repository: https://github.com/ueki-tomohiro/stackdriver_dart
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.4
-  http_extensions: ^1.0.0
-  dio: ^4.0.6
-  stack_trace: ^1.10.0
+  http: ^1.1.0
+  pretty_http_logger: ^1.0.2
+  dio: ^5.3.3
+  stack_trace: ^1.11.0
 
 dev_dependencies:
   flutter_test:
