@@ -1,15 +1,14 @@
 import 'package:http/http.dart';
-import 'package:http_extensions/http_extensions.dart';
 import 'package:stackdriver_dart/stackdriver_dart.dart';
+import 'package:pretty_http_logger/pretty_http_logger.dart';
 
 class HttpApiClient {
   late final StackDriverErrorReporter reporter;
-  late final ExtendedClient _client;
+  late final HttpClientWithMiddleware _client;
 
   HttpApiClient(this.reporter) {
-    _client = ExtendedClient(
-      inner: Client() as BaseClient,
-      extensions: [
+    _client = HttpClientWithMiddleware.build(
+      middlewares: [
         StackDriverReportExtension(reporter: reporter),
       ],
     );
