@@ -6,7 +6,7 @@ import 'dart:io';
 import 'package:dio/dio.dart' as dio;
 import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
-import 'package:http_extensions/http_extensions.dart';
+import 'package:pretty_http_logger/pretty_http_logger.dart';
 import 'package:stack_trace/stack_trace.dart';
 
 part 'dio_extension.dart';
